+++ tools/rescue/Cargo.toml
@@ -41,10 +41,6 @@ diem-api-types = { workspace = true }
 diem-temppath = { workspace = true }
 diem-forge = { workspace = true }
 diem-logger = { workspace = true }
-
 libra-smoke-tests = { workspace = true }
-
-
 regex = { workspace = true }
 smoke-test = { workspace = true }
-
