+++ tools/rescue/tests/rescue_smoke.rs
@@ -1,10 +1,11 @@
+mod support;
+
 use diem_api_types::ViewRequest;
 use diem_config::config::{InitialSafetyRulesConfig, NodeConfig};
-use diem_forge::{LocalNode, Node, NodeExt, SwarmExt, Validator};
+use diem_forge::{LocalNode, NodeExt, SwarmExt, Validator};
 use diem_logger::prelude::*;
 use diem_temppath::TempPath;
-use diem_types::{account_address::AccountAddress, transaction::Transaction};
-use libra_framework::framework_cli::make_template_files;
+use diem_types::transaction::Transaction;
 use libra_smoke_tests::libra_smoke::LibraSmoke;
 use rescue::{diem_db_bootstrapper::BootstrapOpts, rescue_tx::RescueTxOpts};
 use smoke_test::test_utils::{
@@ -13,7 +14,6 @@ use smoke_test::test_utils::{
 };
 use std::{
     fs,
-    path::PathBuf,
     process::Command,
     time::{Duration, Instant},
 };
@@ -52,48 +52,6 @@ async fn wait_for_node(validator: &mut dyn Validator, expected_to_connect: usize
 }
 
 #[tokio::test]
-/// Testing sync only
-async fn test_sync_only_mode() {
-    let num_nodes: usize = 5;
-
-    let mut s = LibraSmoke::new(Some(num_nodes as u8))
-        .await
-        .expect("could not start libra smoke");
-
-    let env = &mut s.swarm;
-
-    println!("1. Set sync_only = true for the last node and check it can sync to others");
-    let node = env.validators_mut().nth(4).unwrap();
-    let mut new_config = node.config().clone();
-    new_config.consensus.sync_only = true;
-    update_node_config_restart(node, new_config.clone());
-    wait_for_node(node, num_nodes - 1).await;
-    // wait for some versions
-    env.wait_for_all_nodes_to_catchup_to_version(10, Duration::from_secs(MAX_CATCH_UP_WAIT_SECS))
-        .await
-        .expect("could not catc up to version 10");
-
-    println!("2. Set sync_only = true for all nodes and restart");
-    for node in env.validators_mut() {
-        let mut node_config = node.config().clone();
-        node_config.consensus.sync_only = true;
-        update_node_config_restart(node, node_config);
-        wait_for_node(node, num_nodes - 1).await;
-    }
-
-    println!("3. delete one node's db and test they can still sync when sync_only is true for every nodes");
-    let node = env.validators_mut().nth(3).unwrap();
-    node.stop();
-    node.clear_storage().await.unwrap();
-    node.start().expect("could not restart node");
-
-    println!("4. verify all nodes are at the same round and no progress being made");
-    env.wait_for_all_nodes_to_catchup(Duration::from_secs(MAX_CATCH_UP_WAIT_SECS))
-        .await
-        .unwrap();
-}
-
-#[tokio::test]
 /// This test verifies the flow of a genesis transaction after the chain starts.
 /// 1. Test the consensus sync_only mode, every node should stop at the same version.
 /// 2. Test the db-bootstrapper applying a manual genesis transaction (remove validator 0) on diemdb directly
@@ -108,7 +66,7 @@ async fn test_genesis_transaction_flow() {
 
     let client = s.client().clone();
 
-    let env = &mut s.swarm;
+    let env: &mut diem_forge::LocalSwarm = &mut s.swarm;
 
     env.wait_for_all_nodes_to_catchup_to_version(10, Duration::from_secs(MAX_CATCH_UP_WAIT_SECS))
         .await
@@ -142,7 +100,7 @@ async fn test_genesis_transaction_flow() {
         .get_peer_id()
         .unwrap();
 
-    let script_path = make_script(first_validator_address);
+    let script_path = support::make_script(first_validator_address);
     dbg!(&script_path);
     assert!(script_path.exists());
 
@@ -230,46 +188,4 @@ async fn test_genesis_transaction_flow() {
 
     node.start().unwrap();
     wait_for_node(node, num_nodes - 2).await;
-}
-
-fn make_script(first_validator_address: AccountAddress) -> PathBuf {
-    let script = format!(
-        r#"
-        script {{
-            use diem_framework::stake;
-            use diem_framework::diem_governance;
-            use diem_framework::block;
-
-            fun main(vm_signer: &signer, framework_signer: &signer) {{
-                stake::remove_validators(framework_signer, &vector[@0x{:?}]);
-                block::emit_writeset_block_event(vm_signer, @0x1);
-                diem_governance::reconfigure(framework_signer);
-            }}
-    }}
-    "#,
-        first_validator_address
-    );
-
-    let framework_path = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .join("..")
-        .join("..")
-        .join("framework")
-        .join("libra-framework");
-
-    let mut temp_script_path = TempPath::new();
-    temp_script_path.create_as_dir().unwrap();
-    temp_script_path.persist();
-
-    dbg!(&temp_script_path.path());
-    assert!(temp_script_path.path().exists());
-
-    make_template_files(
-        temp_script_path.path(),
-        &framework_path,
-        "rescue",
-        Some(script),
-    )
-    .unwrap();
-
-    temp_script_path.path().to_owned()
 }
